Change price variables from double to decimal as precision is needed on the data types

+++ Sales_Tax/Sales_Tax/Models/TaxableItemCategory.cs
@@ -3,6 +3,6 @@
    public class TaxableItemCategory
     {
         public string CategoryName { get; set; }
-        public int CategoryTaxRate { get; set; }
+        public decimal CategoryTaxRate { get; set; }
     }
 }
